Hardcode command and server install list
* Created two new json files `commands.json` & `game_servers.json` to store the
  command list and the server install list.
  - Parsing that json data to populate /controls & /install pages rather than
    gather that data via rce.
  - Reasons for hardcoding:
    - Simplifies project.
    - Less rce calls.
    - Less verification / sanitization.
    - Speeds up initial page loads.
    - Enables removing large chunks of bad code!
* Used configparse to setup basic settings stored on disk in main.conf.
  - Rn only settings are for text color and delete files on web-lgsm entry del.
  - Created working settings page to view / change those settings.

+++ app/views.py
@@ -8,6 +8,7 @@ from werkzeug.security import generate_password_hash
 from .models import *
 from . import db
 from .utils import *
+import configparser
 
 views = Blueprint("views", __name__)
 
@@ -26,52 +27,16 @@ def home():
     return render_template("home.html", user=current_user, installed_servers=server_names)
 
 
-# Controls route helper function. Uses script_path to gather list of short &
-# long cmds and then commits them to the db. Helps speed up page load times if
-# cmds are pulled from db rather than gathered each time via a popen call.
-def add_cmd_to_db(server_name, script_path, disabled_cmds):
-
-    # Requires further validation!!!
-    ## TRY TO VALIDATE EVERYTHING!!!
-    try:
-        stdout = os.popen(script_path).read()
-    except:
-        # For debug.
-        print(sys.exc_info()[0])
-
-    cmds = escape_ansi(stdout)
-
-    long_cmds = []
-    short_cmds = []
-    descriptions = []
-
-    before_commands_line = True
-
-    for line in cmds.splitlines():
-        # Only the command lines have a | char.
-        if "|" not in line:
-            continue
-
-        args = line.split()
-
-        if args[1] in disabled_cmds:
-            continue
-
-        long_cmds.append(args[0])
-        short_cmds.append(args[1])
-        descriptions.append(line.split("|")[1])
-
-    controls = ControlSet(install_name=server_name, short_cmds=','.join(short_cmds), \
-                 long_cmds=','.join(long_cmds), descriptions=','.join(descriptions)) 
-    db.session.add(controls)
-    db.session.commit()
-    return controls
-
 ######### Controls Page #########
 
 @views.route("/controls", methods=['GET'])
 @login_required
 def controls():
+    # Import config data.
+    config = configparser.ConfigParser()
+    config.read('main.conf')
+    text_color = config['aesthetic']['text_color']
+
     # Import meta data.
     meta_data = MetaData.query.get(1)
     base_dir = meta_data.app_install_path
@@ -96,28 +61,6 @@ def controls():
         return redirect(url_for('views.home'))
 
     script_path = server.install_path + '/' + server.script_name
-    disabled_cmds = ['d', 'sk', 'i', 'ai', 'dev']
-
-    controls = ControlSet.query.filter_by(install_name=server_name).first()
-    if controls == None:
-        controls = add_cmd_to_db(server_name, script_path, disabled_cmds)
-
-    class CmdDescriptor:
-        def __init__(self):
-            self.long_cmd  = ""
-            self.short_cmd = ""
-            self.description = ""
-
-    commands = []
-    cmd_strings = zip(controls.short_cmds.split(','), controls.long_cmds.split(','), \
-                                                controls.descriptions.split(','))
-
-    for short_cmd, long_cmd, description in cmd_strings:
-        cmd = CmdDescriptor()
-        cmd.long_cmd = long_cmd
-        cmd.short_cmd = short_cmd
-        cmd.description = description
-        commands.append(cmd)
 
     # Check for hax injectypoo attempt!
     for input_item in (server_name, script_arg):
@@ -127,20 +70,16 @@ def controls():
             return redirect(url_for('views.controls', server=server_name))
 
     if script_arg != None:
-        if script_arg in disabled_cmds:
-            flash("Option disabled", category="error")
-            return redirect(url_for('views.controls', server=server_name))
-
         # Console option, use tmux capture-pane.
-        elif script_arg == "c":
+        if script_arg == "c":
             cmd = f'/usr/bin/tmux capture-pane -pS -5 -t {server.script_name}'
-            return Response(read_process(server.install_path, base_dir, cmd, ""), mimetype= 'text/html')
+            return Response(read_process(server.install_path, base_dir, cmd, text_color, ""), mimetype= 'text/html')
         else:
             cmd = f'{script_path} {script_arg}'
-            return Response(read_process(server.install_path, base_dir, cmd, ""), mimetype= 'text/html')
+            return Response(read_process(server.install_path, base_dir, cmd, text_color, ""), mimetype= 'text/html')
        
     return render_template("controls.html", user=current_user, server_name=server_name, \
-                        server_commands=commands, cmd=script_arg, doggo_img=get_doggo())
+                server_commands=get_commands(), cmd=script_arg, text_color=text_color)
 
 ######### Iframe Default Page #########
 
@@ -149,43 +88,17 @@ def controls():
 def no_output():
     return render_template("no_output.html", user=current_user)
 
-# Install route helper function. Uses the main linuxgsm.sh to gather list of
-# short & long server names and then commits them to the db. Helps speed up
-# page load times if server names are pulled from db rather than gathered each
-# time via a popen call.
-def add_server_list_to_db():
-    try:
-        stdout = os.popen('./linuxgsm.sh list').read()
-    except:
-        # For debug.
-        print(sys.exc_info()[0])
-
-    servers_list = escape_ansi(stdout)
-    short_server_names = []
-    long_server_names = []
-
-    for line in servers_list.splitlines():
-        server_script_name = line.split()[0]
-        server_full_name = line.split(" ", 1)[1].strip()
-
-        # Part of status stdout, not part of listing so ignore.
-        if server_script_name == "fetching":
-            continue
-
-        short_server_names.append(server_script_name)
-        long_server_names.append(server_full_name)
-
-    servers = InstallServer(short_names=','.join(short_server_names), \
-                            long_names=','.join(long_server_names))
-    db.session.add(servers)
-    db.session.commit()
-    return servers
 
 ######### Install Page #########
 
 @views.route("/install", methods=['GET', 'POST'])
 @login_required
 def install():
+    # Import config data.
+    config = configparser.ConfigParser()
+    config.read('main.conf')
+    text_color = config['aesthetic']['text_color']
+
     # Import meta data.
     meta_data = MetaData.query.get(1)
     base_dir = meta_data.app_install_path
@@ -253,32 +166,43 @@ def install():
 
         flash("Game server added!")
 
-        return Response(read_process(install_path, base_dir, setup_cmd, "install"), mimetype= 'text/html')
+        return Response(read_process(install_path, base_dir, setup_cmd, \
+                        text_color, "install"), mimetype= 'text/html')
 
-    servers_list = InstallServer.query.get(1)
-    if servers_list == None:
-        # Rename at some point.
-        servers_list = add_server_list_to_db()
-
-    servers = {}
-    server_name_strings = zip(servers_list.short_names.split(','), \
-                                servers_list.long_names.split(','))
-
-    for short_names, long_names in server_name_strings:
-        servers[short_names] = long_names
-
-    return render_template("install.html", user=current_user, servers=servers, \
-                                        output=output, doggo_img=get_doggo())
+    return render_template("install.html", user=current_user, \
+        servers=get_servers(), output=output, text_color=text_color)
 
 ######### Settings Page #########
 
 @views.route("/settings", methods=['GET', 'POST'])
 @login_required
 def settings():
+    # Import config data.
+    config = configparser.ConfigParser()
+    config.read('main.conf')
+    text_color = config['aesthetic']['text_color']
+    remove_files = config['settings'].getboolean('remove_files')
+
+    if request.method == 'POST':
+        color_pref = request.form.get("text_color")
+        file_pref = request.form.get("delete_files")
+
+        config['settings']['remove_files'] = 'yes'
+        if file_pref == "false":
+            config['settings']['remove_files'] = 'no'
+        
+        config['aesthetic']['text_color'] = text_color
+        if color_pref != None:
+            config['aesthetic']['text_color'] = color_pref
+
+        with open('main.conf', 'w') as configfile:
+             config.write(configfile)
+
+        flash("Settings Updated!")
+        return redirect(url_for('views.settings'))
     
-    flash("Redirecting to controls")
-    return redirect(url_for('views.controls', server="Minecraft", command="m"))
-   # return render_template("settings.html", user=current_user)
+    return render_template("settings.html", user=current_user, \
+                text_color=text_color, remove_files=remove_files)
 
 ######### Add Page #########
 
@@ -319,17 +243,16 @@ def add():
     return render_template("add.html", user=current_user)
 
 # Helper function for del route.
-def del_server(server_name):
+def del_server(server_name, remove_files):
     server = GameServer.query.filter_by(install_name=server_name).first()
     install_path = server.install_path 
 
     GameServer.query.filter_by(install_name=server_name).delete()
-    ControlSet.query.filter_by(install_name=server_name).delete()
     db.session.commit()
 
-# Disabled auto delete rn. Will add configparse option to disable l8t3r.
-    if os.path.isdir(install_path):
-        shutil.rmtree(install_path)
+    if remove_files:
+        if os.path.isdir(install_path):
+            shutil.rmtree(install_path)
     
     flash(f'Game server, {server_name} deleted!')
     return
@@ -339,6 +262,11 @@ def del_server(server_name):
 @views.route("/delete", methods=['GET', 'POST'])
 @login_required
 def delete():
+    # Import config data.
+    config = configparser.ConfigParser()
+    config.read('main.conf')
+    remove_files = config['settings']['remove_files']
+
     # For multiple deletions from home page.
     if request.method == 'POST':
         for server, server_name in request.form.items():
@@ -346,13 +274,13 @@ def delete():
             if contains_bad_chars(server) or contains_bad_chars(server_name):
                 flash("Illegal Character Detected! Stop Haxing!", category="error")
                 return redirect(url_for('views.home'))
-            del_server(server_name)
+            del_server(server_name, remove_files)
     else:
         server_name = request.args.get("server")
         # Always check all input!
         if contains_bad_chars(server_name):
             flash("Illegal Character Detected! Stop Haxing!", category="error")
             return redirect(url_for('views.home'))
-        del_server(server_name)
+        del_server(server_name, remove_files)
 
     return redirect(url_for('views.home'))
